Fix all 67 actionable clippy findings + make CI clippy strict
Result: cargo clippy with the project's silenced-pedantic gate is
now warning-free, and the CI step's continue-on-error is removed
so any new finding fails the build.

Auto-fixed by `cargo clippy --fix` (~46 sites):
- Redundant closures, unnecessary qualifications, redundant
  references, etc. across database.rs, perspective/, gpu.rs.

Manual fixes:

Dead code:
- Removed PerspectiveCommandResult::error (unused since the
  command-error refactor)
- Removed unused imports in database.rs (image::GenericImageView,
  CREATE_NO_WINDOW const) and database/editor.rs (tauri::Manager)
- Removed dead `filename` binding in config.rs::copy_watermark_to_app_data
- #[allow(dead_code)] on Property.completed (deserialised but not
  read), perspective::detection::LineSegment endpoint fields
  (debug-only), perspective::detection::HistogramPeak.angle_deg
  (debug-only), perspective::straighten::LineSegment angle fields
  (debug-only), image_editor::ImageCache.preview_size (cache
  invalidation hook for the future)

&PathBuf -> &Path migration on 5 helper signatures:
- database::scan_folder_for_properties
- database::is_valid_property_folder
- database::create_property_folder_structure (now uses
  to_path_buf() to feed the spawn_blocking 'static closure)
- database::watermark::copy_and_process_folder_with_config
- perspective::commands::process_single_image

Type aliases for two too-complex tuple types:
- database.rs ScanRow (7-tuple of fs scan results)
- straighten.rs IndexedLine<'a> (Vec<(usize, &HoughLine)>)

Identical-block collapse in perspective/detection.rs (the two
180-degree branches collapsed to a single arm).

#[allow]-with-rationale on three hot-path lints that would cost
correctness/clarity to "fix":
- gpu::ImageProcessor::rotate_and_adjust + gpu::gpu_rotate_and_adjust
  (clippy::too_many_arguments — 8 independent adjustment params,
  bundling into a struct just shifts the cost)
- gpu::cpu_clahe (clippy::needless_range_loop — multi-axis
  tile_mappings[ty][tx][i] indexing and cdf accumulation don't
  translate cleanly to iterators)

CI workflow: dropped `continue-on-error: true` from the clippy step
and renamed it from "informational" to "deny warnings". From here
on a new actionable clippy finding fails CI.

cargo test:    128 passing
cargo clippy:  0 errors, 0 warnings (with the project gate)

+++ src-tauri/src/gpu.rs
@@ -150,6 +150,10 @@ impl ImageProcessor {
     /// Uploads the image once, dispatches rotation, feeds output directly
     /// into adjustments, downloads final result once.
     /// Saves ~20-30ms per full-res image by eliminating 2 redundant PCIe transfers.
+    ///
+    /// Each adjustment parameter is independent — bundling them into a struct
+    /// would just shift the complexity, so we silence the lint here.
+    #[allow(clippy::too_many_arguments)]
     pub fn rotate_and_adjust(
         &self,
         img: &DynamicImage,
@@ -1074,6 +1078,9 @@ fn gpu_watermark_blend(
 // GPU Fused Rotation + Adjustments (single round-trip)
 // ============================================================================
 
+// Mirrors `ImageProcessor::rotate_and_adjust` — independent params, struct
+// bundling would just push complexity around.
+#[allow(clippy::too_many_arguments)]
 fn gpu_rotate_and_adjust(
     ctx: &GpuContext,
     img: &DynamicImage,
@@ -1948,7 +1955,7 @@ fn cpu_gradient_histogram(
 
     let w = width as usize;
     let h = height as usize;
-    let num_bins = GRADIENT_HISTOGRAM_BINS as usize;
+    let num_bins = GRADIENT_HISTOGRAM_BINS;
 
     // Each thread accumulates a local histogram, then we reduce
     let local_hists: Vec<Vec<f64>> = (1..h - 1)
@@ -2083,6 +2090,11 @@ fn cpu_bilateral(gray_pixels: &[u8], width: u32, height: u32) -> Vec<u8> {
 
 /// CPU fallback for CLAHE on grayscale data.
 /// Uses 8x8 grid, clip limit 2.0 (matching the GPU shader defaults).
+//
+// Multi-axis indexing into `tile_mappings[ty][tx][i]` and the
+// `cdf[i] = cdf[i-1] + hist[i]` accumulation don't translate
+// cleanly into iterator chains, so silence needless_range_loop.
+#[allow(clippy::needless_range_loop)]
 fn cpu_clahe(gray_pixels: &[u8], width: u32, height: u32) -> Vec<u8> {
     let w = width as usize;
     let h = height as usize;
